pyproject.toml
[tool.poetry]
name = "mt-server"
version = "0.1.0"
description = ""
authors = ["hiiggs <yurasik@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.3"
django-rest-framework = "^0.1.0"
pillow = "^10.3.0"
drf-spectacular = "^0.27.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
